requirements.txt
streamlit
pandas
pillow
fpdf
openpyxl
st-gsheets-connection